feat: Archiving mutations (#11)
* feat: Archiving mutations

+++ app/graphql/mutations/delete_rig.rb
@@ -1,59 +1,64 @@
 # frozen_string_literal: true
 
 module Mutations
-  class DeleteTicketType < Mutations::BaseMutation
-    field :ticket_type, Types::TicketTypeType, null: true
+  class DeleteRig < Mutations::BaseMutation
+    field :rig, Types::RigType, null: true
     field :errors, [String], null: true
     field :field_errors, [Types::FieldErrorType], null: true
 
     argument :id, Int, required: true
 
     def resolve(id:)
-      model = TicketType.find(id)
+      model = Rig.find(id)
 
-      # Flag as deleted if any slots use this ticket type
-      if model.slots.empty?
+      # Flag as deleted if any slots use this rig
+      # or if this rig has been inspected
+      if model.slots.empty? && model.rig_inspections.empty?
         model.destroy
       else
         model.discard
       end
 
       {
-        ticket_type: model.reload,
+        rig: model.reload,
         field_errors: nil,
         errors: nil
       }
     rescue ActiveRecord::RecordInvalid => invalid
       # Failed save, return the errors to the client
       {
-        ticket_type: nil,
+        rig: nil,
         field_errors: invalid.record.errors.messages.map { |field, messages| { field: field, message: messages.first } },
         errors: invalid.record.errors.full_messages
       }
     rescue ActiveRecord::RecordNotSaved => error
       # Failed save, return the errors to the client
       {
-        ticket_type: nil,
+        rig: nil,
         field_errors: nil,
         errors: error.record.errors.full_messages
       }
     rescue ActiveRecord::RecordNotFound => error
       {
-        ticket_type: nil,
+        rig: nil,
         field_errors: nil,
         errors: [ error.message ]
       }
     end
 
     def authorized?(id: nil, attributes: nil)
-      dz_user = Dropzone.find(id).dropzone_users.find_by(user_id: context[:current_user])
-      if dz_user.can? :deleteDropzone
-        true
-      else
-        return false, {
-          errors: ["You cant delete this dropzone"]
-        }
-      end
+      rig = Rig.find(id)
+      dropzone_ids = rig.user.dropzone_users.pluck(:dropzone_id)
+
+      return true if rig.user_id == context[:current_user].id
+      return true if dropzone_ids.count == 1 && context[:current_resource].can?(
+        :deleteUserRig,
+        dropzone_id: dropzone_ids.first
+      )
+
+      return false, {
+        errors: ["You cant delete this rig"]
+      }
     end
   end
 end
